+++ src/components/Products/Product/styles.js
@@ -28,6 +28,7 @@ export default makeStyles((theme) => ({
     display: 'flex',
     justifyContent: 'space-between',
     alignItems: 'center',
+    flexWrap: 'wrap',
   },
   cardContent: {
     height: '100%',
